rename Modal to EventDetailsModal
