+++ frontend/src/components/EventList/CalendarDay/Event/EventDetailsModal/EventDetailsModal_styles.js
